requirements.txt
h5py==2.8.0
matplotlib==2.2.3
numpy==1.14.5
scipy==1.1.0
tensorflow-gpu==1.5.1
tensorflow-tensorboard==1.5.1
tqdm==4.40.2
